Update project_clean
Spotted issue where first input was not taken in consideration when running the code. Addressed the issue where get_user_guess() was called twice in the main function and the first call was ignored in the check_guess function. 
Modified the code to store the initial user guess in a variable and use that variable consistently: